-Updated MainMenu class by added backgroundMusic thread into it, which lets music starts when menu is loaded and stop when menu is exited
